+++ src/Rect.java
@@ -1,6 +1,6 @@
 	public class Rect {
-		int x,y,w,h;
-		public Rect(int x, int y, int w, int h) {this.x = x; this.y = y; this.w = w; this.h = h;}
+		float x,y,w,h;
+		public Rect(float x, float y, float w, float h) {this.x = x; this.y = y; this.w = w; this.h = h;}
 		
 		public Boolean Contains(Particle p) 
 		{
